+++ ydb/core/tx/columnshard/engines/reader/simple_reader/iterator/sys_view/constructor.cpp
@@ -1,0 +1,5 @@
+#include "constructor.h"
+
+namespace NKikimr::NOlap::NReader::NSimple {
+
+}   // namespace NKikimr::NOlap
