+++ ydb/core/tx/columnshard/engines/reader/simple_reader/iterator/sys_view/source.cpp
@@ -1,0 +1,56 @@
+#include "source.h"
+
+#include <ydb/core/tx/conveyor_composite/usage/service.h>
+
+namespace NKikimr::NOlap::NReader::NSimple {
+namespace {
+class TPortionAccessorFetchingSubscriber: public IDataAccessorRequestsSubscriber {
+private:
+    NReader::NCommon::TFetchingScriptCursor Step;
+    std::shared_ptr<NReader::NSimple::IDataSource> Source;
+    const NColumnShard::TCounterGuard Guard;
+    virtual const std::shared_ptr<const TAtomicCounter>& DoGetAbortionFlag() const override {
+        return Source->GetContext()->GetCommonContext()->GetAbortionFlag();
+    }
+
+    virtual void DoOnRequestsFinished(TDataAccessorsResult&& result) override {
+        FOR_DEBUG_LOG(NKikimrServices::COLUMNSHARD_SCAN_EVLOG, Source->AddEvent("facc"));
+        if (result.HasErrors()) {
+            Source->GetContext()->GetCommonContext()->AbortWithError("has errors on portion accessors restore");
+            return;
+        }
+        AFL_VERIFY(result.GetPortions().size() == 1)("count", result.GetPortions().size());
+        Source->MutableStageData().SetPortionAccessor(std::move(result.ExtractPortionsVector().front()));
+        Source->InitUsedRawBytes();
+        AFL_VERIFY(Step.Next());
+        const auto& commonContext = *Source->GetContext()->GetCommonContext();
+        auto task = std::make_shared<NReader::NCommon::TStepAction>(Source, std::move(Step), commonContext.GetScanActorId(), false);
+        NConveyorComposite::TScanServiceOperator::SendTaskToExecute(task, commonContext.GetConveyorProcessId());
+    }
+
+public:
+    TPortionAccessorFetchingSubscriber(
+        const NReader::NCommon::TFetchingScriptCursor& step, const std::shared_ptr<NReader::NSimple::IDataSource>& source)
+        : Step(step)
+        , Source(source)
+        , Guard(Source->GetContext()->GetCommonContext()->GetCounters().GetFetcherAcessorsGuard()) {
+    }
+};
+
+}   // namespace
+
+bool TSysViewPortionChunksInfo::DoStartFetchingAccessor(
+    const std::shared_ptr<IDataSource>& sourcePtr, const NReader::NCommon::TFetchingScriptCursor& step) {
+    AFL_VERIFY(!GetStageData().HasPortionAccessor());
+    AFL_DEBUG(NKikimrServices::TX_COLUMNSHARD_SCAN)("event", step.GetName())("fetching_info", step.DebugString());
+
+    std::shared_ptr<TDataAccessorsRequest> request =
+        std::make_shared<TDataAccessorsRequest>(NGeneralCache::TPortionsMetadataCachePolicy::EConsumer::SCAN);
+    request->AddPortion(GetPortion());
+    request->SetColumnIds(GetContext()->GetAllUsageColumns()->GetColumnIds());
+    request->RegisterSubscriber(std::make_shared<TPortionAccessorFetchingSubscriber>(step, sourcePtr));
+    GetContext()->GetCommonContext()->GetDataAccessorsManager()->AskData(request);
+    return true;
+}
+
+}   // namespace NKikimr::NOlap::NReader::NSimple
