+++ ydb/core/tx/columnshard/engines/reader/simple_reader/iterator/sys_view/ya.make
@@ -1,0 +1,12 @@
+LIBRARY()
+
+SRCS(
+    source.cpp
+    constructor.cpp
+)
+
+PEERDIR(
+    ydb/core/tx/columnshard/engines/reader/common_reader/iterator
+)
+
+END()
