+++ ydb/core/tx/columnshard/engines/reader/simple_reader/iterator/ya.make
@@ -16,6 +16,7 @@ PEERDIR(
     ydb/core/tx/columnshard/engines/reader/common_reader/iterator
     ydb/core/tx/columnshard/engines/reader/simple_reader/iterator/collections
     ydb/core/tx/columnshard/engines/reader/simple_reader/iterator/sync_points
+    ydb/core/tx/columnshard/engines/reader/simple_reader/iterator/sys_view
     ydb/core/tx/columnshard/engines/reader/simple_reader/duplicates
     ydb/core/tx/conveyor/usage
     ydb/core/tx/limiter/grouped_memory/usage
